+++ src/main/java/com/innowise/taxi/entity/Car.java
@@ -1,19 +1,19 @@
 package com.innowise.taxi.entity;
 
 public class Car {
-  private Long id;
+  private int id;
   private String model;
   private String plateNumber;
   private int year;
 
-  public Car(Long id, String model, String plateNumber, int year) {
+  public Car(int id, String model, String plateNumber, int year) {
     this.id = id;
     this.model = model;
     this.plateNumber = plateNumber;
     this.year = year;
   }
 
-  public Long getId() {
+  public int getId() {
     return id;
   }
